lesson 9

+++ api/src/main/java/ru/gb/wintermarket/api/dto/CartDto.java
@@ -1,9 +1,10 @@
 package ru.gb.wintermarket.api.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 public class CartDto {
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public List<CartItemDto> getItems() {
         return items;
@@ -13,11 +14,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
